refactor: Deduplicate antagonistic review interfaces and inline logic
