+++ apps/server/src/services/lead-engineer-processors.ts
@@ -7,7 +7,7 @@
 
 import { createLogger } from '@protolabsai/utils';
 import { resolveModelString, DEFAULT_MODELS } from '@protolabsai/model-resolver';
-import { areDependenciesSatisfied } from '@protolabsai/dependency-resolver';
+import { getBlockingDependencies } from '@protolabsai/dependency-resolver';
 import type { AgentRole, Feature } from '@protolabsai/types';
 import { getWorkflowSettings, getPhaseModelWithOverrides } from '../lib/settings-helpers.js';
 import { simpleQuery } from '../providers/simple-query-service.js';
@@ -41,17 +41,9 @@ export class IntakeProcessor implements StateProcessor {
     // Validate dependencies using the shared resolver (same logic as loadPendingFeatures)
     if (feature.dependencies && feature.dependencies.length > 0) {
       const allFeatures = await this.serviceContext.featureLoader.getAll(ctx.projectPath);
+      const unmetDeps = getBlockingDependencies(feature, allFeatures);
 
-      if (!areDependenciesSatisfied(feature, allFeatures)) {
-        const unmetDeps = feature.dependencies.filter((depId) => {
-          const dep = allFeatures.find((f) => f.id === depId);
-          if (!dep) return true;
-          if (dep.isFoundation) {
-            return dep.status !== 'done' && dep.status !== 'completed' && dep.status !== 'verified';
-          }
-          return !dep.status || !['completed', 'verified', 'done', 'review'].includes(dep.status);
-        });
-
+      if (unmetDeps.length > 0) {
         ctx.escalationReason = `Unmet dependencies: ${unmetDeps.join(', ')}`;
         logger.warn(`[INTAKE] Feature has ${unmetDeps.length} unmet dependencies`, {
           featureId: feature.id,
